+++ src/shim/kmq/pcidev.h
@@ -12,7 +12,7 @@ namespace shim_xdna {
 class pdev_kmq : public pdev
 {
 public:
-  using shim_xdna::pdev::pdev;
+  using pdev::pdev;
 
 public:
   bool
