+++ src/shim/pcidev.h
@@ -4,7 +4,7 @@
 #ifndef PCIDEV_XDNA_H
 #define PCIDEV_XDNA_H
 
-#include "pcidrv.h"
+#include "platform.h"
 #include "core/pcie/linux/pcidev.h"
 
 namespace shim_xdna {
@@ -12,7 +12,7 @@ namespace shim_xdna {
 class pdev : public xrt_core::pci::dev
 {
 public:
-  pdev(std::shared_ptr<const drv>& driver, const std::string& sysfs_name);
+  pdev(std::shared_ptr<const platform_drv>& driver, const std::string& sysfs_name);
   ~pdev();
 
   xrt_core::device::handle_type
@@ -47,24 +47,15 @@ public:
   is_umq() const = 0;
 
 private:
-  class dev_fd {
-  public:
-    dev_fd(int fd) : m_fd(fd) {}
-    ~dev_fd() { ::close(m_fd); }
-    int get() { return m_fd; }
-  private:
-    int m_fd = -1;
-  };
   virtual void
   on_first_open() const = 0;
 
   virtual void
   on_last_close() const = 0;
 
-  mutable std::unique_ptr<dev_fd> m_dev_fd;
   mutable int m_dev_users = 0;
   mutable std::mutex m_lock;
-  std::shared_ptr<const drv> m_driver;
+  std::shared_ptr<const platform_drv> m_driver;
 };
 
 }
