+++ src/shim/pcidrv.cpp
@@ -2,9 +2,9 @@
 // Copyright (C) 2022-2025, Advanced Micro Devices, Inc. All rights reserved.
 
 #include "pcidrv.h"
-#include "kmq/pcidev.h"
-#include "umq/pcidev.h"
 #include "shim_debug.h"
+#include <sys/types.h>
+#include <dirent.h>
 
 namespace shim_xdna {
 
@@ -15,20 +15,23 @@ is_user() const
   return true;
 }
 
-std::shared_ptr<xrt_core::pci::dev>
+std::string
 drv::
-create_pcidev(const std::string& sysfs) const
+get_dev_node(const std::string& sysfs_name) const
 {
-  auto driver = std::dynamic_pointer_cast<const drv>(shared_from_this());
+  const std::string sysfs_root{"/sys/bus/pci/devices/"};
+  const std::string dev_path_dir = sysfs_root + sysfs_name + "/" + sysfs_dev_node_dir();
+  const auto prefix = dev_node_prefix();
 
-  if (m_device_type == AMDXDNA_DEV_TYPE_UNKNOWN)
-    m_device_type = get_dev_type(sysfs);
-
-  if (m_device_type == AMDXDNA_DEV_TYPE_KMQ)
-    return std::make_shared<pdev_kmq>(driver, sysfs);
-  if (m_device_type == AMDXDNA_DEV_TYPE_UMQ)
-    return std::make_shared<pdev_umq>(driver, sysfs);
-  shim_err(EINVAL, "Unknown device type: %d", m_device_type);
+  auto dp = opendir(dev_path_dir.c_str());
+  if (dp) {
+    while (auto entry = readdir(dp)) {
+      std::string dirname{entry->d_name};
+      if(dirname.compare(0, prefix.size(), prefix) == 0)
+        return std::string("/dev/") + dev_node_dir() + "/" + dirname;
+    }
+  }
+  shim_err(EINVAL, "Bad sysfs name: %s", sysfs_name.c_str());
 }
 
 }
